+++ dag_graph_validator/src/data_struct/treeLib.h
@@ -20,7 +20,7 @@ extern "C"
     */
     typedef struct TreeNode
     {
-        int data;
+        unsigned long data;
         struct TreeNode *left;
         struct TreeNode *right;
     } TreeNodeType;
@@ -73,7 +73,7 @@ extern "C"
         \param value The value to look for
         \return True if the value is in the tree, false otherwise
      */
-    bool treeContains(const TreeType *self, int value);
+    bool treeContains(const TreeType *self, unsigned long value);
 
     /*!
         \brief Insert a value in the tree and return false if the value was already present
@@ -81,7 +81,7 @@ extern "C"
         \param value The value to insert
         \return True if the value was inserted, false otherwise
      */
-    bool treeInsert(TreeType *self, int value);
+    bool treeInsert(TreeType *self, unsigned long value);
 
     /*!
         \brief Remove a value from the tree and return false if the value was not present
@@ -89,14 +89,14 @@ extern "C"
         \param value The value to remove
         \return True if the value was removed, false otherwise
      */
-    bool treeRemove(TreeType *self, int value);
+    bool treeRemove(TreeType *self, unsigned long value);
 
     /*!
         \brief Function type to walk in the tree
         \param value The value of the node
         \param user_data The data to pass to the function
      */
-    typedef void (*treeFuncT)(int value, void *userData);
+    typedef void (*treeFuncT)(unsigned long value, void *userData);
 
     /*!
         \brief Walk in the tree in pre order and call the function with user_data as a second argument
